refactor(varcon): Pull out core types
