+++ crates/varcon-core/src/parser.rs
